% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{argument_lut}
\alias{argument_lut}
\title{Provides a convenient look up table to decifer command line arguments}
\usage{
argument_lut()
}
\value{
tibble of common argument descriptions
\itemize{
\item{-b "YYYY-MM-DD HH:MM", Specify the begin (start) time for predictions.}
\item{-e "YYYY-MM-DD HH:MM", Specify the end (stop) time for predictions.}
\item{-f "t", Specify output format as a table as CSV}
\item{-m "r", Specify mode to be raw}
\item{-o "/path/to/output", Redirect output to the specified file (appends).}
\item{-s "HH:MM", Specify the step interval, in hours and minutes, for raw mode predictions.}
\item{-v "", Print version string and exit.}
}
}
\description{
Provides a convenient look up table to decifer command line arguments
}
