% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fetch_locations}
\alias{fetch_locations}
\title{Retrieve an online listing of stations}
\usage{
fetch_locations(uri = "https://flaterco.com/xtide/locations.html",
  decimal = TRUE, pattern = NA, ...)
}
\arguments{
\item{uri}{character, url of the online locations listing}

\item{decimal}{logical, if TRUE convert Latitude and Longtitude to decimal}

\item{pattern}{character, a pattern to match in station name - possibly a
regular expression or a fixed pattern. Passed to \code{base::grepl()}}

\item{...}{further arguments for \code{base::grepl()} such as \code{fixed}}
}
\value{
table of site locations
}
\description{
Retrieve an online listing of stations
}
