% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{get_plain_mode}
\alias{get_plain_mode}
\title{Retrieve the tide events for a specified location}
\usage{
get_plain_mode(location = "Portland", time = c(Sys.time(), Sys.time() +
  4 * (24 * 3600)), ofile = tempfile(), compress = FALSE,
  app = Sys.which("/opt/bin/tide"))
}
\arguments{
\item{location}{the name of the location}

\item{time}{2 element vector of POSIXct time to start and stop start <= time < stop.
Defaults to [now, now + 4 days]}

\item{ofile}{character, the output filename. Defaults to \code{tempfile()}}

\item{compress}{logical, if TRUE compress the file. Defaults to TRUE.}

\item{app}{the name of the application. Defaults to \code{Sys.which("/opt/bin/tide")}}
}
\value{
list with
\itemize{
   \item{location, the name of the location}
   \item{file, the name of the output file}
   \item{moon, tibble of moon rise/set, possible with 0 rows}
   \item{sun, tibble of sun rise/set, possibly with 0 rows}
   \item{tide, tibble of tide height and stage, possibly with 0 rows}
}
}
\description{
Retrieve the tide events for a specified location
}
