% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{latlon_as_decimal}
\alias{latlon_as_decimal}
\title{Convert longitude and latitude to decimal equivalents}
\usage{
latlon_as_decimal(x, sep = c(",", NA)[1])
}
\arguments{
\item{x}{character, vector of one or two elements in the form of
"43.6233째 N, 70.2067째 W" or ["43.6233째 N", "70.2067째 W"]}

\item{sep}{character or NA, if character the input is split
if NA we assume that a two element vector is provided}
}
\value{
a two element numeric of [lat, lon]
}
\description{
Convert longitude and latitude to decimal equivalents
}
