% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_plain_mode}
\alias{read_plain_mode}
\title{Read plain mode data}
\usage{
read_plain_mode(filename, tz = Sys.timezone())
}
\arguments{
\item{filename}{character or connection}

\item{tz}{character, the time zone for the location and time requested}
}
\value{
list with
\itemize{
   \item{location, the name of the location}
   \item{lonlat, numeric two element vector of [lon,lat]}
   \item{file, the name of the output file}
   \item{status, status code from calling the function - either zero (sucess) or non-zero}
   \item{moon, tibble of moon rise/set, possible with 0 rows}
   \item{sun, tibble of sun rise/set, possibly with 0 rows}
   \item{tide, tibble of tide height and stage, possibly with 0 rows}
}
}
\description{
Read plain mode data
}
