% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{get_raw_mode}
\alias{get_raw_mode}
\title{Retrieve the raw tide values for a specified location}
\usage{
get_raw_mode(location = "Portland", time = c(Sys.time(), Sys.time() + 4
  * (24 * 3600)), step = "01:00", ofile = tempfile(),
  compress = FALSE, app = Sys.which("/opt/bin/tide"))
}
\arguments{
\item{location}{the name of the location}

\item{time}{2 element vector of POSIXct time to start and stop start <= time < stop.
Defaults to [now, now + 4 days]}

\item{step}{character, time step as 'HH:MM'. Defaults tp "01:00"}

\item{ofile}{character, the output filename. Defaults to \code{tempfile()}}

\item{compress}{logical, if TRUE compress the file. Defaults to TRUE.}

\item{app}{the name of the application. Defaults to \code{Sys.which("/opt/bin/tide")}}
}
\value{
list with
\itemize{
   \item{location, the name of the location}
   \item{file, the name of the output file}
   \item{data, tibble of data, possible with 0 rows}
}
}
\description{
Retrieve the raw tide values for a specified location
}
