% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_raw_mode}
\alias{read_raw_mode}
\title{Read raw mode data}
\usage{
read_raw_mode(filename, tz = "UTC")
}
\arguments{
\item{filename}{character or connection}

\item{tz}{character, the time zone for the location and time requested}
}
\value{
list with
\itemize{
   \item{file, the name of the output file}
   \item{data, tibble of data, possible with 0 rows}
}
}
\description{
Read raw mode data
}
